Add example of Auto Command for Vim in README.adoc
An example of the actual command needed to make it work.
If the command is not correct, please let me know. It took me a lot of hunting around just to get that syntax highlighting working for adoc documents. This example might help people deal with the problem swiftly (since this repo almost always appears in search results).

Hope it helps.